+++ collection/collection.go
@@ -1,8 +1,11 @@
 package collection
 
+import "gorm.io/gorm"
+
 // Collection is a struct that stores a list of HQs
 type Collection struct {
-	ID        int    `json:"id"`
+	gorm.Model
+	// ID        int    `json:"id" gorm:"primaryKey"`
 	Name      string `json:"name"`
 	Publisher string `json:"published"`
 }
