+++ collection/collection.server.go
@@ -4,15 +4,17 @@ import (
 	"encoding/json"
 	"log"
 	"net/http"
-	"strconv"
 
 	"github.com/gorilla/mux"
+	"hq-collections.com/database"
 )
 
 func HandleCollections(w http.ResponseWriter, r *http.Request) {
 	switch r.Method {
 	case http.MethodGet:
-		collectionList := getCollectionsList()
+		var collectionList []Collection
+		database.DB.Find(&collectionList)
+
 		j, err := json.Marshal(collectionList)
 		if err != nil {
 			log.Fatal(err)
@@ -22,20 +24,20 @@ func HandleCollections(w http.ResponseWriter, r *http.Request) {
 			log.Fatal(err)
 		}
 	case http.MethodPost:
-		var collection Collection
-		err := json.NewDecoder(r.Body).Decode(&collection)
-		if err != nil {
-			log.Print(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		_, err = addOrUpdateCollection(collection)
-		if err != nil {
-			log.Print(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		w.WriteHeader(http.StatusCreated)
+		// var collection Collection
+		// err := json.NewDecoder(r.Body).Decode(&collection)
+		// if err != nil {
+		// 	log.Print(err)
+		// 	w.WriteHeader(http.StatusBadRequest)
+		// 	return
+		// }
+		// _, err = addOrUpdateCollection(collection)
+		// if err != nil {
+		// 	log.Print(err)
+		// 	w.WriteHeader(http.StatusBadRequest)
+		// 	return
+		// }
+		// w.WriteHeader(http.StatusCreated)
 	case http.MethodOptions:
 		return
 	default:
@@ -49,50 +51,50 @@ func HandleCollection(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
-	collectionID, err := strconv.Atoi(params["id"])
-	if err != nil {
-		log.Print(err)
-		w.WriteHeader(http.StatusNotFound)
-		return
-	}
+	// collectionID, err := strconv.Atoi(params["id"])
+	// if err != nil {
+	// 	log.Print(err)
+	// 	w.WriteHeader(http.StatusNotFound)
+	// 	return
+	// }
 	switch r.Method {
 	case http.MethodGet:
-		collection := getCollection(collectionID)
-		if collection == nil {
-			w.WriteHeader(http.StatusNotFound)
-			return
-		}
-		j, err := json.Marshal(collection)
-		if err != nil {
-			log.Print(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		_, err = w.Write(j)
-		if err != nil {
-			log.Fatal(err)
-		}
+		// collection := getCollection(collectionID)
+		// if collection == nil {
+		// 	w.WriteHeader(http.StatusNotFound)
+		// 	return
+		// }
+		// j, err := json.Marshal(collection)
+		// if err != nil {
+		// 	log.Print(err)
+		// 	w.WriteHeader(http.StatusBadRequest)
+		// 	return
+		// }
+		// _, err = w.Write(j)
+		// if err != nil {
+		// 	log.Fatal(err)
+		// }
 
 	case http.MethodPut:
-		var collection Collection
-		err := json.NewDecoder(r.Body).Decode(&collection)
-		if err != nil {
-			log.Print(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		if collection.ID != collectionID {
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		_, err = addOrUpdateCollection(collection)
-		if err != nil {
-			log.Print(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
+		// var collection Collection
+		// err := json.NewDecoder(r.Body).Decode(&collection)
+		// if err != nil {
+		// 	log.Print(err)
+		// 	w.WriteHeader(http.StatusBadRequest)
+		// 	return
+		// }
+		// if collection.ID != collectionID {
+		// 	w.WriteHeader(http.StatusBadRequest)
+		// 	return
+		// }
+		// _, err = addOrUpdateCollection(collection)
+		// if err != nil {
+		// 	log.Print(err)
+		// 	w.WriteHeader(http.StatusBadRequest)
+		// 	return
+		// }
 	case http.MethodDelete:
-		removeCollection(collectionID)
+		// removeCollection(collectionID)
 
 	case http.MethodOptions:
 		return
@@ -107,29 +109,29 @@ func HandleCollectionVolumes(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
-	collectionID, err := strconv.Atoi(params["id"])
-	if err != nil {
-		log.Print(err)
-		w.WriteHeader(http.StatusNotFound)
-		return
-	}
+	// collectionID, err := strconv.Atoi(params["id"])
+	// if err != nil {
+	// 	log.Print(err)
+	// 	w.WriteHeader(http.StatusNotFound)
+	// 	return
+	// }
 	switch r.Method {
 	case http.MethodGet:
-		collection := getCollection(collectionID)
-		if collection == nil {
-			w.WriteHeader(http.StatusNotFound)
-			return
-		}
-		j, err := json.Marshal(collection.getCollectionVolumes())
-		if err != nil {
-			log.Print(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		_, err = w.Write(j)
-		if err != nil {
-			log.Fatal(err)
-		}
+		// collection := getCollection(collectionID)
+		// if collection == nil {
+		// 	w.WriteHeader(http.StatusNotFound)
+		// 	return
+		// }
+		// j, err := json.Marshal(collection.getCollectionVolumes())
+		// if err != nil {
+		// 	log.Print(err)
+		// 	w.WriteHeader(http.StatusBadRequest)
+		// 	return
+		// }
+		// _, err = w.Write(j)
+		// if err != nil {
+		// 	log.Fatal(err)
+		// }
 	case http.MethodOptions:
 		return
 	default:
