+++ go.mod
@@ -4,7 +4,8 @@ go 1.13
 
 require (
 	github.com/gorilla/mux v1.8.0
-	github.com/jinzhu/gorm v1.9.16 // indirect
+	github.com/jinzhu/gorm v1.9.16
 	github.com/joho/godotenv v1.3.0 // indirect
-	gorm.io/gorm v1.20.12 // indirect
+	gorm.io/driver/postgres v1.0.8
+	gorm.io/gorm v1.20.12
 )
