+++ routes.go
@@ -6,7 +6,6 @@ import (
 
 	"hq-collections.com/collection"
 	"hq-collections.com/cors"
-	"hq-collections.com/volume"
 
 	"github.com/gorilla/mux"
 )
@@ -14,10 +13,10 @@ import (
 // SetupRoutes builds the routes for the API
 func SetupRoutes(apiBasePath string) {
 	collectionsHandler := http.HandlerFunc(collection.HandleCollections)
-	collectionHandler := http.HandlerFunc(collection.HandleCollection)
-	collectionVolumesHandler := http.HandlerFunc(collection.HandleCollectionVolumes)
-	volumesHandler := http.HandlerFunc(volume.HandleVolumes)
-	volumeHandler := http.HandlerFunc(volume.HandleVolume)
+	// collectionHandler := http.HandlerFunc(collection.HandleCollection)
+	// collectionVolumesHandler := http.HandlerFunc(collection.HandleCollectionVolumes)
+	// volumesHandler := http.HandlerFunc(volume.HandleVolumes)
+	// volumeHandler := http.HandlerFunc(volume.HandleVolume)
 
 	router := mux.NewRouter()
 	router.Use(cors.Middleware)
@@ -25,12 +24,12 @@ func SetupRoutes(apiBasePath string) {
 		Methods("GET", "POST", "PUT", "DELETE", "OPTIONS").
 		Subrouter()
 
-	apiRouter.HandleFunc("/collections/{id}", collectionHandler)
-	apiRouter.HandleFunc("/collections/{id}/volumes", collectionVolumesHandler)
+	// apiRouter.HandleFunc("/collections/{id}", collectionHandler)
+	// apiRouter.HandleFunc("/collections/{id}/volumes", collectionVolumesHandler)
 	apiRouter.HandleFunc("/collections", collectionsHandler)
 
-	apiRouter.HandleFunc("/volumes/{id}", volumeHandler)
-	apiRouter.HandleFunc("/volumes", volumesHandler)
+	// apiRouter.HandleFunc("/volumes/{id}", volumeHandler)
+	// apiRouter.HandleFunc("/volumes", volumesHandler)
 
 	http.Handle("/", router)
